+++ src/session.rs
@@ -3,11 +3,10 @@
 use crate::{
     control_code::ControlCode,
     error::Error,
-    log,
     needle::Needle,
-    process::{Process, Stream},
-    stream::TryStream,
-    Found, Expect,
+    process::{NonBlocking, Process, Stream},
+    stream::{empty, log, stream::TryStream},
+    Expect, Found,
 };
 use std::{
     convert::TryInto,
@@ -16,13 +15,10 @@ use std::{
     time::{self, Duration},
 };
 
-#[cfg(all(unix, feature = "async"))]
-use futures_lite::AsyncWriteExt;
-
 /// Session represents a spawned process and its streams.
 /// It controlls process and communication with it.
 #[derive(Debug)]
-pub struct Session<P, S: Stream> {
+pub struct Session<P, S> {
     proc: P,
     stream: TryStream<S>,
     expect_timeout: Option<Duration>,
@@ -40,6 +36,13 @@ where
     }
 }
 
+impl<P, S> Session<P, S> {
+    /// Set the pty session's expect timeout.
+    pub fn set_expect_timeout(&mut self, expect_timeout: Option<Duration>) {
+        self.expect_timeout = expect_timeout;
+    }
+}
+
 impl<P, S: Stream> Session<P, S> {
     /// Create a session
     pub fn new(process: P, stream: S) -> io::Result<Self> {
@@ -51,36 +54,18 @@ impl<P, S: Stream> Session<P, S> {
         })
     }
 
-    /// Set the pty session's expect timeout.
-    pub fn set_expect_timeout(&mut self, expect_timeout: Option<Duration>) {
-        self.expect_timeout = expect_timeout;
-    }
-
-    #[cfg(not(feature = "async"))]
     pub fn with_log<W: Write>(
         self,
         logger: W,
     ) -> Result<Session<P, log::LoggedStream<S, W>>, Error> {
-        let (session, old) = self.swap_stream(log::EmptyStream)?;
+        let (session, old) = self.swap_stream(empty::EmptyStream)?;
         let stream = log::LoggedStream::new(old, logger);
         let (session, _) = session.swap_stream(stream)?;
 
         Ok(session)
     }
 
-    #[cfg(feature = "async")]
-    pub fn with_log<W: Write + Unpin>(
-        self,
-        logger: W,
-    ) -> Result<Session<log::LoggedStream<S, W>>, Error> {
-        let (session, old) = self.swap_stream(log::EmptyStream)?;
-        let stream = log::LoggedStream::new(old, logger);
-        let (session, _) = session.swap_stream(stream)?;
-
-        Ok(session)
-    }
-
-    pub(crate) fn swap_stream<N: Stream>(self, stream: N) -> io::Result<(Session<P, N>, S)> {
+    fn swap_stream<N: Stream>(self, stream: N) -> io::Result<(Session<P, N>, S)> {
         let (stream, old) = self.stream.swap_stream(stream)?;
         Ok((
             Session {
@@ -228,8 +213,7 @@ impl<P, S: Stream> Expect for Session<P, S> {
     }
 }
 
-#[cfg(not(feature = "async"))]
-impl<P, S: Stream> Session<P, S> {
+impl<P, S: Read + NonBlocking> Session<P, S> {
     /// Try to read in a non-blocking mode.
     ///
     /// Returns `[std::io::ErrorKind::WouldBlock]`
@@ -244,125 +228,7 @@ impl<P, S: Stream> Session<P, S> {
     }
 }
 
-#[cfg(not(feature = "async"))]
-impl<P: Process, S: Stream> Session<P, S> {
-    /// Send `EOF` indicator to a child process.
-    ///
-    /// Often `eof` char handled as it would be a CTRL-C.
-    pub fn send_eof(&mut self) -> io::Result<()> {
-        self.stream.write_all(&[self.proc.get_eof_char()?])
-    }
-
-    /// Send `INTR` indicator to a child process.
-    ///
-    /// Often `intr` char handled as it would be a CTRL-D.
-    pub fn send_intr(&mut self) -> io::Result<()> {
-        self.stream.write_all(&[self.proc.get_intr_char()?])
-    }
-}
-
-#[cfg(all(feature = "async", not(windows)))]
-impl<P: Process + Unpin, S: Stream> Session<P, S> {
-    /// Send text to child's `STDIN`.
-    ///
-    /// To write bytes you can use a [std::io::Write] operations instead.
-    pub async fn send(&mut self, s: impl AsRef<str>) -> io::Result<()> {
-        self.stream.write_all(s.as_ref().as_bytes()).await
-    }
-
-    /// Send a line to child's `STDIN`.
-    pub async fn send_line(&mut self, s: impl AsRef<str>) -> io::Result<()> {
-        #[cfg(windows)]
-        const LINE_ENDING: &[u8] = b"\r\n";
-        #[cfg(not(windows))]
-        const LINE_ENDING: &[u8] = b"\n";
-
-        let _ = self.write_all(s.as_ref().as_bytes()).await?;
-        let _ = self.write_all(LINE_ENDING).await?;
-        self.flush().await?;
-
-        Ok(())
-    }
-
-    /// Send controll character to a child process.
-    ///
-    /// You must be carefull passing a char or &str as an argument.
-    /// If you pass an unexpected controll you'll get a error.
-    /// So it may be better to use [ControlCode].
-    ///
-    /// ```no_run
-    /// use expectrl::{Session, ControlCode};
-    /// use std::process::Command;
-    ///
-    /// # futures_lite::future::block_on(async {
-    /// let mut process = Session::spawn(Command::new("cat")).unwrap();
-    /// process.send_control(ControlCode::EndOfText).await.unwrap(); // sends CTRL^C
-    /// process.send_control('C').await.unwrap(); // sends CTRL^C
-    /// process.send_control("^C").await.unwrap(); // sends CTRL^C
-    /// # });
-    /// ```
-    pub async fn send_control(&mut self, code: impl TryInto<ControlCode>) -> io::Result<()> {
-        let code = code.try_into().map_err(|_| {
-            io::Error::new(io::ErrorKind::Other, "Failed to parse a control character")
-        })?;
-        self.stream.write_all(&[code.into()]).await
-    }
-
-    /// Send `EOF` indicator to a child process.
-    ///
-    /// Often `eof` char handled as it would be a CTRL-C.
-    pub async fn send_eof(&mut self) -> io::Result<()> {
-        self.stream.write_all(&[self.proc.get_eof_char()?]).await
-    }
-
-    /// Send `INTR` indicator to a child process.
-    ///
-    /// Often `intr` char handled as it would be a CTRL-D.
-    pub async fn send_intr(&mut self) -> io::Result<()> {
-        self.stream.write_all(&[self.proc.get_intr_char()?]).await
-    }
-
-    pub(crate) fn get_available(&mut self) -> &[u8] {
-        self.stream.get_available()
-    }
-
-    pub(crate) fn consume_from_buffer(&mut self, n: usize) {
-        self.stream.consume_from_buffer(n);
-    }
-}
-
-impl<P, S: Stream> Deref for Session<P, S> {
-    type Target = P;
-
-    fn deref(&self) -> &Self::Target {
-        &self.proc
-    }
-}
-
-impl<P, S: Stream> DerefMut for Session<P, S> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.proc
-    }
-}
-
-#[cfg(feature = "async")]
-impl<P, S: Stream> Session<P, S> {
-    /// Try to read in a non-blocking mode.
-    ///
-    /// Returns `[std::io::ErrorKind::WouldBlock]`
-    /// in case if there's nothing to read.
-    pub async fn try_read(&mut self, buf: &mut [u8]) -> io::Result<usize> {
-        self.stream.try_read(buf).await
-    }
-
-    /// Verifyes if stream is empty or not.
-    pub async fn is_empty(&mut self) -> io::Result<bool> {
-        self.stream.is_empty().await
-    }
-}
-
-#[cfg(not(feature = "async"))]
-impl<P, S: Stream> std::io::Write for Session<P, S> {
+impl<P, S: Write> Write for Session<P, S> {
     fn write(&mut self, buf: &[u8]) -> std::io::Result<usize> {
         self.stream.write(buf)
     }
@@ -376,15 +242,13 @@ impl<P, S: Stream> std::io::Write for Session<P, S> {
     }
 }
 
-#[cfg(not(feature = "async"))]
-impl<P, S: Stream> std::io::Read for Session<P, S> {
+impl<P, S: Read> Read for Session<P, S> {
     fn read(&mut self, buf: &mut [u8]) -> std::io::Result<usize> {
         self.stream.read(buf)
     }
 }
 
-#[cfg(not(feature = "async"))]
-impl<P, S: Stream> std::io::BufRead for Session<P, S> {
+impl<P, S: Read> BufRead for Session<P, S> {
     fn fill_buf(&mut self) -> std::io::Result<&[u8]> {
         self.stream.fill_buf()
     }
@@ -394,52 +258,16 @@ impl<P, S: Stream> std::io::BufRead for Session<P, S> {
     }
 }
 
-#[cfg(feature = "async")]
-impl<P: Unpin, S: Stream> futures_lite::io::AsyncWrite for Session<P, S> {
-    fn poll_write(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &[u8],
-    ) -> std::task::Poll<std::io::Result<usize>> {
-        std::pin::Pin::new(&mut self.get_mut().stream).poll_write(cx, buf)
-    }
+impl<P, S> Deref for Session<P, S> {
+    type Target = P;
 
-    fn poll_flush(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::io::Result<()>> {
-        std::pin::Pin::new(&mut self.get_mut().stream).poll_flush(cx)
-    }
-
-    fn poll_close(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::io::Result<()>> {
-        std::pin::Pin::new(&mut self.get_mut().stream).poll_close(cx)
+    fn deref(&self) -> &Self::Target {
+        &self.proc
     }
 }
 
-#[cfg(feature = "async")]
-impl<P: Unpin, S: Stream> futures_lite::io::AsyncRead for Session<P, S> {
-    fn poll_read(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-        buf: &mut [u8],
-    ) -> std::task::Poll<std::io::Result<usize>> {
-        std::pin::Pin::new(&mut self.get_mut().stream).poll_read(cx, buf)
-    }
-}
-
-#[cfg(feature = "async")]
-impl<P: Unpin, S: Stream> futures_lite::io::AsyncBufRead for Session<P, S> {
-    fn poll_fill_buf(
-        self: std::pin::Pin<&mut Self>,
-        cx: &mut std::task::Context<'_>,
-    ) -> std::task::Poll<std::io::Result<&[u8]>> {
-        std::pin::Pin::new(&mut self.get_mut().stream).poll_fill_buf(cx)
-    }
-
-    fn consume(mut self: std::pin::Pin<&mut Self>, amt: usize) {
-        std::pin::Pin::new(&mut self.stream).consume(amt);
+impl<P, S> DerefMut for Session<P, S> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.proc
     }
 }
