+++ tests/io.rs
@@ -1,4 +1,4 @@
-use expectrl::{session::Session, ControlCode, Expect, PtySession, Stream};
+use expectrl::{process::Stream, session::Session, ControlCode, Expect};
 use std::{thread, time::Duration};
 
 #[cfg(unix)]
@@ -22,7 +22,7 @@ use std::io::{BufRead, Read, Write};
 #[test]
 #[cfg(unix)]
 fn send_controll() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
     _p_send_control(&mut proc, ControlCode::EOT).unwrap();
     assert_eq!(proc.wait().unwrap(), WaitStatus::Exited(proc.pid(), 0),);
 }
@@ -31,8 +31,7 @@ fn send_controll() {
 #[cfg(windows)]
 fn send_controll() {
     let mut proc =
-        PtySession::spawn_command(ProcAttr::cmd("powershell -C ping localhost".to_string()))
-            .unwrap();
+        Session::spawn_command(ProcAttr::cmd("powershell -C ping localhost".to_string())).unwrap();
 
     // give powershell a bit time
     thread::sleep(Duration::from_millis(100));
@@ -47,7 +46,7 @@ fn send_controll() {
 #[test]
 #[cfg(unix)]
 fn send() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
     _p_send(&mut proc, "hello cat\n").unwrap();
 
     // give cat a time to react on input
@@ -63,10 +62,9 @@ fn send() {
 #[test]
 #[cfg(windows)]
 fn send() {
-    let mut proc = PtySession::spawn_command(
-        ProcAttr::default().commandline("powershell -C type".to_string()),
-    )
-    .unwrap();
+    let mut proc =
+        Session::spawn_command(ProcAttr::default().commandline("powershell -C type".to_string()))
+            .unwrap();
     thread::sleep(Duration::from_millis(1000));
 
     _p_send(&mut proc, "hello cat\r\n").unwrap();
@@ -91,7 +89,7 @@ fn send() {
 #[test]
 #[cfg(unix)]
 fn send_line() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     _p_send_line(&mut proc, "hello cat").unwrap();
 
@@ -108,8 +106,7 @@ fn send_line() {
 #[test]
 #[cfg(windows)]
 fn send_line() {
-    let mut proc =
-        PtySession::spawn_command(ProcAttr::cmd("powershell -C type".to_string())).unwrap();
+    let mut proc = Session::spawn_command(ProcAttr::cmd("powershell -C type".to_string())).unwrap();
 
     thread::sleep(Duration::from_millis(1000));
     _p_send_line(&mut proc, "hello cat").unwrap();
@@ -130,7 +127,7 @@ fn send_line() {
 #[test]
 #[cfg(unix)]
 fn try_read_by_byte() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     assert_eq!(
         _p_try_read(&mut proc, &mut [0; 1]).unwrap_err().kind(),
@@ -166,8 +163,7 @@ fn try_read_by_byte() {
     // Mustn't it be turned down?
 
     let mut proc =
-        PtySession::spawn_command(ProcAttr::default().commandline("powershell".to_string()))
-            .unwrap();
+        Session::spawn_command(ProcAttr::default().commandline("powershell".to_string())).unwrap();
     _p_send_line(
         &mut proc,
         "while (1) { read-host | set r; if (!$r) { break }}",
@@ -196,7 +192,7 @@ fn try_read_by_byte() {
 #[test]
 #[cfg(unix)]
 fn blocking_read_after_non_blocking() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     assert!(_p_is_empty(&mut proc).unwrap());
 
@@ -227,8 +223,7 @@ fn blocking_read_after_non_blocking() {
 #[cfg(windows)]
 fn blocking_read_after_non_blocking() {
     let mut proc =
-        PtySession::spawn_command(ProcAttr::default().commandline("powershell".to_string()))
-            .unwrap();
+        Session::spawn_command(ProcAttr::default().commandline("powershell".to_string())).unwrap();
     _p_send_line(
         &mut proc,
         "while (1) { read-host | set r; if (!$r) { break }}",
@@ -254,7 +249,7 @@ fn blocking_read_after_non_blocking() {
 #[test]
 #[cfg(unix)]
 fn try_read() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     let mut buf = vec![0; 128];
     assert_eq!(
@@ -279,8 +274,7 @@ fn try_read() {
 #[cfg(windows)]
 fn try_read() {
     let mut proc =
-        PtySession::spawn_command(ProcAttr::default().commandline("powershell".to_string()))
-            .unwrap();
+        Session::spawn_command(ProcAttr::default().commandline("powershell".to_string())).unwrap();
     _p_send_line(
         &mut proc,
         "while (1) { read-host | set r; if (!$r) { break }}",
@@ -311,7 +305,7 @@ fn try_read() {
 #[test]
 #[cfg(unix)]
 fn blocking_read_after_non_blocking_try_read() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     let mut buf = vec![0; 1];
     assert_eq!(
@@ -344,8 +338,7 @@ fn blocking_read_after_non_blocking_try_read() {
 #[test]
 #[cfg(windows)]
 fn blocking_read_after_non_blocking_try_read() {
-    let mut proc =
-        PtySession::spawn_command(ProcAttr::cmd("powershell -C type".to_string())).unwrap();
+    let mut proc = Session::spawn_command(ProcAttr::cmd("powershell -C type".to_string())).unwrap();
 
     thread::sleep(Duration::from_millis(1000));
     while !_p_try_read(&mut proc, &mut [0; 1]).is_err() {}
@@ -371,7 +364,7 @@ fn blocking_read_after_non_blocking_try_read() {
 #[cfg(unix)]
 #[test]
 fn try_read_after_eof() {
-    let mut proc = PtySession::spawn_command(Command::new("cat")).unwrap();
+    let mut proc = Session::spawn_command(Command::new("cat")).unwrap();
 
     _p_send_line(&mut proc, "hello").unwrap();
 
@@ -392,7 +385,7 @@ fn try_read_after_eof() {
 fn try_read_after_process_exit() {
     let mut command = Command::new("echo");
     command.arg("hello cat");
-    let mut proc = PtySession::spawn_command(command).unwrap();
+    let mut proc = Session::spawn_command(command).unwrap();
 
     assert_eq!(proc.wait().unwrap(), WaitStatus::Exited(proc.pid(), 0));
 
@@ -418,7 +411,7 @@ fn try_read_after_process_exit() {
 #[test]
 #[cfg(windows)]
 fn try_read_after_process_exit() {
-    let mut proc = PtySession::spawn_command(ProcAttr::cmd("echo hello cat".to_string())).unwrap();
+    let mut proc = Session::spawn_command(ProcAttr::cmd("echo hello cat".to_string())).unwrap();
 
     assert_eq!(proc.wait(None).unwrap(), 0);
 
@@ -435,7 +428,7 @@ fn try_read_after_process_exit() {
 fn try_read_to_end() {
     let mut cmd = Command::new("echo");
     cmd.arg("Hello World");
-    let mut proc = PtySession::spawn_command(cmd).unwrap();
+    let mut proc = Session::spawn_command(cmd).unwrap();
 
     let mut buf: Vec<u8> = Vec::new();
     loop {
@@ -499,7 +492,7 @@ fn continues_try_reads() {
 #[cfg(not(target_os = "macos"))]
 #[cfg(not(windows))]
 fn automatic_stop_of_interact() {
-    let mut p = PtySession::spawn_command(Command::new("ls")).unwrap();
+    let mut p = Session::spawn_command(Command::new("ls")).unwrap();
     let status = _p_interact(&mut p).unwrap();
 
     // It may be finished not only because process is done but
@@ -510,7 +503,7 @@ fn automatic_stop_of_interact() {
     ));
 
     // check that second spawn works
-    let mut p = PtySession::spawn_command(Command::new("ls")).unwrap();
+    let mut p = Session::spawn_command(Command::new("ls")).unwrap();
     let status = _p_interact(&mut p).unwrap();
     assert!(matches!(
         status,
@@ -522,14 +515,14 @@ fn automatic_stop_of_interact() {
 #[cfg(not(target_os = "macos"))]
 #[cfg(not(windows))]
 fn spawn_after_interact() {
-    let mut p = PtySession::spawn_command(Command::new("ls")).unwrap();
+    let mut p = Session::spawn_command(Command::new("ls")).unwrap();
     let _ = _p_interact(&mut p).unwrap();
 
-    let p = PtySession::spawn_command(Command::new("ls")).unwrap();
+    let p = Session::spawn_command(Command::new("ls")).unwrap();
     assert!(matches!(p.wait().unwrap(), WaitStatus::Exited(_, 0)));
 }
 
-fn _p_read(mut proc: impl Expect, buf: &mut [u8]) -> std::io::Result<usize> {
+fn _p_read(mut proc: impl Read, buf: &mut [u8]) -> std::io::Result<usize> {
     #[cfg(not(feature = "async"))]
     {
         proc.read(buf)
@@ -540,7 +533,7 @@ fn _p_read(mut proc: impl Expect, buf: &mut [u8]) -> std::io::Result<usize> {
     }
 }
 
-fn _p_write_all<S: Stream>(proc: &mut impl Expect, buf: &[u8]) -> std::io::Result<()> {
+fn _p_write_all(proc: &mut impl Write, buf: &[u8]) -> std::io::Result<()> {
     #[cfg(not(feature = "async"))]
     {
         proc.write_all(buf)
@@ -551,7 +544,7 @@ fn _p_write_all<S: Stream>(proc: &mut impl Expect, buf: &[u8]) -> std::io::Resul
     }
 }
 
-fn _p_flush(proc: &mut impl Expect) -> std::io::Result<()> {
+fn _p_flush(proc: &mut impl Write) -> std::io::Result<()> {
     #[cfg(not(feature = "async"))]
     {
         proc.flush()
@@ -595,7 +588,7 @@ fn _p_send_control(proc: &mut impl Expect, buf: impl Into<ControlCode>) -> std::
     }
 }
 
-fn _p_read_to_string(proc: &mut impl Expect) -> std::io::Result<String> {
+fn _p_read_to_string(proc: &mut impl BufRead) -> std::io::Result<String> {
     let mut buf = String::new();
     #[cfg(not(feature = "async"))]
     {
@@ -608,7 +601,7 @@ fn _p_read_to_string(proc: &mut impl Expect) -> std::io::Result<String> {
     Ok(buf)
 }
 
-fn _p_read_to_end(proc: &mut PtySession) -> std::io::Result<Vec<u8>> {
+fn _p_read_to_end<P, S: Read>(proc: &mut Session<P, S>) -> std::io::Result<Vec<u8>> {
     let mut buf = Vec::new();
     #[cfg(not(feature = "async"))]
     {
@@ -621,7 +614,7 @@ fn _p_read_to_end(proc: &mut PtySession) -> std::io::Result<Vec<u8>> {
     Ok(buf)
 }
 
-fn _p_read_until(proc: &mut PtySession, ch: u8) -> std::io::Result<Vec<u8>> {
+fn _p_read_until<P, S: Read>(proc: &mut Session<P, S>, ch: u8) -> std::io::Result<Vec<u8>> {
     let mut buf = Vec::new();
     #[cfg(not(feature = "async"))]
     {
@@ -636,7 +629,7 @@ fn _p_read_until(proc: &mut PtySession, ch: u8) -> std::io::Result<Vec<u8>> {
     Ok(buf)
 }
 
-fn _p_read_line(proc: &mut PtySession) -> std::io::Result<String> {
+fn _p_read_line<P, S: Read>(proc: &mut Session<P, S>) -> std::io::Result<String> {
     let mut buf = String::new();
     #[cfg(not(feature = "async"))]
     {
@@ -649,7 +642,7 @@ fn _p_read_line(proc: &mut PtySession) -> std::io::Result<String> {
     Ok(buf)
 }
 
-fn _p_is_empty(proc: &mut PtySession) -> std::io::Result<bool> {
+fn _p_is_empty<P, S: Stream>(proc: &mut Session<P, S>) -> std::io::Result<bool> {
     #[cfg(not(feature = "async"))]
     {
         proc.is_empty()
@@ -660,7 +653,7 @@ fn _p_is_empty(proc: &mut PtySession) -> std::io::Result<bool> {
     }
 }
 
-fn _p_try_read(proc: &mut PtySession, buf: &mut [u8]) -> std::io::Result<usize> {
+fn _p_try_read<P, S: Stream>(proc: &mut Session<P, S>, buf: &mut [u8]) -> std::io::Result<usize> {
     #[cfg(not(feature = "async"))]
     {
         proc.try_read(buf)
@@ -672,7 +665,9 @@ fn _p_try_read(proc: &mut PtySession, buf: &mut [u8]) -> std::io::Result<usize> 
 }
 
 #[cfg(unix)]
-fn _p_interact(proc: &mut PtySession) -> Result<WaitStatus, expectrl::Error> {
+fn _p_interact(
+    proc: &mut Session<expectrl::process::unix::UnixProcess, expectrl::process::unix::PtyStream>,
+) -> Result<WaitStatus, expectrl::Error> {
     #[cfg(not(feature = "async"))]
     {
         proc.interact()
